@@ -1,8 +1,9 @@
 package com.jeremyliao.lebapp;
 
 import android.content.Intent;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+
+import androidx.test.rule.ActivityTestRule;
+import androidx.test.runner.AndroidJUnit4;
 
 import com.jeremyliao.lebapp.activity.IpcTestActivity;
 import com.jeremyliao.lebapp.activity.TestActivity;
