@@ -1,15 +1,14 @@
 package com.jeremyliao.lebapp.service;
 
 import android.app.Service;
-import android.arch.lifecycle.Observer;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
-import android.util.Log;
 import android.widget.Toast;
 
+import androidx.annotation.Nullable;
+import androidx.lifecycle.Observer;
+
 import com.jeremyliao.lebapp.LiveEventBusDemo;
-import com.jeremyliao.lebapp.app.DemoApplication;
 import com.jeremyliao.liveeventbus.LiveEventBus;
 
 /**
@@ -23,12 +22,6 @@ public class IpcService extends Service {
         LiveEventBus
                 .get(LiveEventBusDemo.KEY_TEST_BROADCAST, String.class)
                 .observeForever(observer);
-        LiveEventBus
-                .get(LiveEventBusDemo.KEY_TEST_BROADCAST_IN_APP, String.class)
-                .observeForever(observer);
-        LiveEventBus
-                .get(LiveEventBusDemo.KEY_TEST_BROADCAST_GLOBAL, String.class)
-                .observeForever(observer);
     }
 
     @Override
@@ -36,12 +29,6 @@ public class IpcService extends Service {
         super.onDestroy();
         LiveEventBus
                 .get(LiveEventBusDemo.KEY_TEST_BROADCAST, String.class)
-                .removeObserver(observer);
-        LiveEventBus
-                .get(LiveEventBusDemo.KEY_TEST_BROADCAST_IN_APP, String.class)
-                .removeObserver(observer);
-        LiveEventBus
-                .get(LiveEventBusDemo.KEY_TEST_BROADCAST_GLOBAL, String.class)
                 .removeObserver(observer);
     }
 
